# services/auth-service/openapi.yaml
openapi: 3.0.3
info:
  title: Auth Service API
  description: Authentication and Authorization service for Project Tanuki
  version: 0.0.1-SNAPSHOT
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/v1/auth/refresh:
    post:
      tags:
        - auth-controller
      summary: Refresh access token
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /api/v1/auth/google/exchange:
    post:
      tags:
        - auth-controller
      summary: Exchange one-time code for AuthResponse
      operationId: exchangeGoogleCode
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - user
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      required:
        - id
        - email
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        createdAt:
          type: string
          format: date-time
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRole'
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
    UserStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - PENDING
        - SUSPENDED
        - DELETED
    UserRole:
      type: string
      enum:
        - USER
        - PAID_USER
        - ADMIN
        - MODERATOR
        - SUPPORT
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT